# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevCourse Finder public API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/v1/catalog/options": {
      "get": {
        "operationId": "catalogOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Options"
                }
              }
            }
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "HTTP method is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "catalogOptions"
      }
    },
    "/api/v1/courses": {
      "get": {
        "operationId": "searchCourses",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "go",
                "python",
                "java",
                "javascript"
              ]
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "basics",
                "backend",
                "frontend",
                "fullstack",
                "automation"
              ]
            }
          },
          {
            "name": "experience",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "none",
                "basics",
                "projects",
                "working",
                "switch",
                "experienced"
              ]
            }
          },
          {
            "name": "goal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "try",
                "job",
                "switch",
                "deepen"
              ]
            }
          },
          {
            "name": "budget",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "free",
                "paid"
              ]
            }
          },
          {
            "name": "support",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "self",
                "review",
                "mentor"
              ]
            }
          },
          {
            "name": "schedule",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "flexible",
                "scheduled"
              ]
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "recent",
                "price_asc",
                "price_desc",
                "duration"
              ]
            }
          },
          {
            "name": "min",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10000000000
            },
            "description": "Amount in kopecks; exact current full price only"
          },
          {
            "name": "max",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10000000000
            },
            "description": "Amount in kopecks; exact current full price only"
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 168
            },
            "description": "Search parameter"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000000
            },
            "description": "Search parameter"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 48
            },
            "description": "Search parameter"
          },
          {
            "name": "include_free",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "include_closed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "HTTP method is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "searchCourses"
      }
    },
    "/api/v1/courses/{slug}": {
      "get": {
        "operationId": "courseDetail",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "HTTP method is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "courseDetail"
      }
    },
    "/api/v1/compare": {
      "get": {
        "operationId": "compareOffers",
        "parameters": [
          {
            "name": "offer_ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[^,]+(,[^,]+)*$"
            },
            "description": "Comma-separated IDs; 1–3 distinct nonempty IDs. Duplicates are ignored while preserving the first occurrence order."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "HTTP method is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "compareOffers"
      }
    },
    "/health/live": {
      "get": {
        "operationId": "liveness",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "HTTP method is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "liveness"
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "readiness",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "HTTP method is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "readiness"
      }
    },
    "/api/v1/events": {
      "post": {
        "operationId": "recordEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Event accepted"
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "HTTP method is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "recordEvent"
      }
    },
    "/out/{offer_id}": {
      "get": {
        "operationId": "outbound",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to approved HTTPS catalog URL; analytics failure does not block redirect",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "HTTP method is not supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "summary": "outbound"
      }
    }
  },
  "components": {
    "schemas": {
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "topics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "audience": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "checked_at": {
            "type": "string",
            "format": "date-time"
          },
          "demo": {
            "type": "boolean"
          },
          "offers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Offer"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "provider",
          "language",
          "direction",
          "summary",
          "source",
          "status",
          "topics",
          "audience",
          "goals",
          "checked_at",
          "demo",
          "offers"
        ]
      },
      "Offer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price_kind": {
            "type": "string"
          },
          "schedule": {
            "type": "string"
          },
          "enrollment": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "price": {
            "type": [
              "integer",
              "null"
            ]
          },
          "hours": {
            "type": [
              "integer",
              "null"
            ]
          },
          "weeks": {
            "type": [
              "integer",
              "null"
            ]
          },
          "free": {
            "type": "boolean"
          },
          "review": {
            "type": "boolean"
          },
          "mentor": {
            "type": "boolean"
          },
          "price_checked_at": {
            "type": "string",
            "format": "date-time"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "price_kind",
          "schedule",
          "enrollment",
          "url",
          "price",
          "hours",
          "weeks",
          "free",
          "review",
          "mentor",
          "price_checked_at"
        ]
      },
      "Result": {
        "type": "object",
        "properties": {
          "course": {
            "$ref": "#/components/schemas/Course"
          },
          "offer": {
            "$ref": "#/components/schemas/Offer"
          },
          "effective_price": {
            "type": [
              "integer",
              "null"
            ]
          },
          "reasons": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "stale": {
            "type": "boolean"
          }
        },
        "required": [
          "course",
          "offer",
          "effective_price",
          "reasons",
          "stale"
        ]
      },
      "Unavailable": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "unavailable": {
            "const": true,
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "enum": [
              "closed"
            ],
            "description": "Present when the tariff exists but enrollment is closed."
          }
        },
        "required": [
          "id",
          "unavailable"
        ]
      },
      "Comparison": {
        "type": "array",
        "items": {
          "oneOf": [
            {
              "$ref": "#/components/schemas/Result"
            },
            {
              "$ref": "#/components/schemas/Unavailable"
            }
          ]
        }
      },
      "CourseDetail": {
        "type": "object",
        "properties": {
          "course": {
            "$ref": "#/components/schemas/Course"
          },
          "offers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Result"
            }
          }
        },
        "required": [
          "course",
          "offers"
        ]
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Result"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "applied_filters": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "data_as_of": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "applied_filters",
          "data_as_of"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "invalid_request",
              "not_found",
              "rate_limited",
              "unavailable",
              "internal_error",
              "method_not_allowed"
            ]
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "code",
          "request_id"
        ]
      },
      "Options": {
        "type": "object",
        "properties": {
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "directions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "languages",
          "directions"
        ]
      },
      "Health": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "kind": {
            "type": "string",
            "enum": [
              "search",
              "empty",
              "view",
              "compare"
            ]
          },
          "course_id": {
            "type": "string",
            "maxLength": 100
          },
          "language": {
            "type": "string",
            "enum": [
              "",
              "go",
              "python",
              "java",
              "javascript"
            ]
          },
          "goal": {
            "type": "string",
            "enum": [
              "",
              "try",
              "job",
              "switch",
              "deepen"
            ]
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          }
        },
        "required": [
          "id",
          "kind"
        ],
        "additionalProperties": false
      }
    }
  },
  "security": []
}